fix uint64 and remove unused impure immutables

+++ service_contracts/src/FilecoinWarmStorageService.sol
@@ -17,7 +17,7 @@ import {Extsload} from "./Extsload.sol";
 
 uint256 constant NO_PROVING_DEADLINE = 0;
 uint256 constant BYTES_PER_LEAF = 32; // Each leaf is 32 bytes
-uint256 constant CHALLENGES_PER_PROOF = 5;
+uint64 constant CHALLENGES_PER_PROOF = 5;
 uint256 constant COMMISSION_MAX_BPS = 10000; // 100% in basis points
 
 /// @title FilecoinWarmStorageService
